Write Ruby setup script
This script installs Bundler if necessary and possible, else points you
to their troubleshooting page to install manually. Once available, it
`bundle install`s our needed gems, which includes a Ruby language
version check, and then runs a dummy test.

Because the purpose of this is to test dependencies, I'm committing the
lock file. I've tested the bundle on versions 2.7.5, 3.0.3, and
3.1.1.