Add more help text for common errors
When on OSX system ruby, which is 2.6.8 and already end-of-lifed,
installing gems required elevated permissions. Using `--user-install`
works smoothly, and we shouldn't need any `$PATH` modification since
the exercise won't be running any gem executables.

When bundler works, it will then request a newer Ruby version, so we
link to Ruby installation instructions.